+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Run the control (baseline) tests in a single test-runner invocation instead of one per test, letting the runner coordinate its own parallelism (https://github.com/svenvanheugten/mutannot/pull/39)
+
 ### Removed
 
 - Remove the `annotate-type` command (https://github.com/svenvanheugten/mutannot/pull/41)